Add animation to the charts and refactor helper and graphics
