+++ src/styles/display-infos/event.css
